Project 2: CalcYouLater
The purpose of this program is to implement a Reverse Polish Notation (RPN) calculator. The program will take in a file or cin input of number, bools, or strings onto a data structure with the functionality of a stack and perform operations on the input. The program will also be able to handle errors in invalid input and print out error messages.