Project 3: Zap
The purpose of this program is to compress and decompress text files using the Huffman coding algorithm. The program can either "zap" or "unzap" a file. The zap command will compress a ASCII text file, compress it using the Huffman coding algorithm, and store the result in a specified output file, as well display the number of bits of the encoded text. The unzap command will take a previously zapped file, decompress it, and store the resulting ASCII text in a specified output file.