Project 4: gerp
The purpose of this program is to create a search tool that allows the user to search for words in a directory of files. The user can search for a word, either case sensitive or case insensitive, and the program will return the line number and file path where the word was found. The user can also search for a word with a specific variation of capitalization using the command "@i" or “@insensitive” followed by a string. The program also allows a user to redirect the output of a search by using the "@f" query command. The program is quit by using the query command “@q” or “@quit”.